server/backend/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["annleefores <annleetestmail@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
kubernetes = "^28.1.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
